+++ src/writer/basic.rs
@@ -15,7 +15,6 @@ use std::{
     cmp,
     fmt::{Debug, Display},
     io,
-    marker::PhantomData,
     str::FromStr,
 };
 
@@ -95,7 +94,7 @@ impl FromStr for Coloring {
 /// [`Scenario`]: gherkin::Scenario
 /// [`writer::Normalized`]: crate::writer::Normalized
 #[derive(Debug, Deref, DerefMut)]
-pub struct Basic<W, Out: io::Write = io::Stdout> {
+pub struct Basic<Out: io::Write = io::Stdout> {
     /// [`io::Write`] implementor to write the output into.
     #[deref]
     #[deref_mut]
@@ -115,16 +114,10 @@ pub struct Basic<W, Out: io::Write = io::Stdout> {
     ///
     /// [1]: gherkin::Step::docstring
     verbose: bool,
-
-    /// Exists for better type inference/resolution, especially when used with
-    /// [`WriterExt`].
-    ///
-    /// [`WriterExt`]: crate::WriterExt
-    _world: PhantomData<W>,
 }
 
 #[async_trait(?Send)]
-impl<W, Out> Writer<W> for Basic<W, Out>
+impl<W, Out> Writer<W> for Basic<Out>
 where
     W: World + Debug,
     Out: io::Write,
@@ -156,7 +149,7 @@ where
 }
 
 #[async_trait(?Send)]
-impl<'val, W, Val, Out> ArbitraryWriter<'val, W, Val> for Basic<W, Out>
+impl<'val, W, Val, Out> ArbitraryWriter<'val, W, Val> for Basic<Out>
 where
     W: World + Debug,
     Val: AsRef<str> + 'val,
@@ -172,13 +165,13 @@ where
     }
 }
 
-impl<W: Debug> Default for Basic<W> {
+impl Default for Basic {
     fn default() -> Self {
         Self::new(io::stdout(), Coloring::Auto, false)
     }
 }
 
-impl<W: Debug, Out: io::Write> Basic<W, Out> {
+impl<Out: io::Write> Basic<Out> {
     /// Creates a new [`Basic`] [`Writer`].
     #[must_use]
     pub fn new(output: Out, color: Coloring, verbose: bool) -> Self {
@@ -188,7 +181,6 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
             indent: 0,
             lines_to_clear: 0,
             verbose: false,
-            _world: PhantomData,
         };
         basic.apply_cli(Cli { verbose, color });
         basic
@@ -248,7 +240,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     /// [scenario]: event::Rule::Scenario
     /// [started]: event::Rule::Started
     /// [`Rule`]: gherkin::Rule
-    pub(crate) fn rule(
+    pub(crate) fn rule<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         rule: &gherkin::Rule,
@@ -294,7 +286,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     /// [started]: event::Scenario::Started
     /// [step]: event::Step
     /// [`Scenario`]: gherkin::Scenario
-    pub(crate) fn scenario(
+    pub(crate) fn scenario<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         scenario: &gherkin::Scenario,
@@ -309,8 +301,8 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
             Scenario::Hook(_, Hook::Started) => {
                 self.indent += 4;
             }
-            Scenario::Hook(which, Hook::Failed(w, info)) => {
-                self.hook_failed(feat, scenario, *which, w.as_deref(), info)?;
+            Scenario::Hook(which, Hook::Failed(world, info)) => {
+                self.hook_failed(feat, scenario, *which, world.as_ref(), info)?;
                 self.indent = self.indent.saturating_sub(4);
             }
             Scenario::Hook(_, Hook::Passed) => {
@@ -331,7 +323,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     ///
     /// [failed]: event::Hook::Failed
     /// [`Scenario`]: gherkin::Scenario
-    pub(crate) fn hook_failed(
+    pub(crate) fn hook_failed<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         sc: &gherkin::Scenario,
@@ -387,7 +379,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     /// [skipped]: event::Step::Skipped
     /// [started]: event::Step::Started
     /// [`Step`]: gherkin::Step
-    pub(crate) fn step(
+    pub(crate) fn step<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         step: &gherkin::Step,
@@ -408,7 +400,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
                 self.indent = self.indent.saturating_sub(4);
             }
             Step::Failed(c, w, i) => {
-                self.step_failed(feat, step, c.as_ref(), w.as_deref(), i)?;
+                self.step_failed(feat, step, c.as_ref(), w.as_ref(), i)?;
                 self.indent = self.indent.saturating_sub(4);
             }
         }
@@ -543,7 +535,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     ///
     /// [failed]: event::Step::Failed
     /// [`Step`]: gherkin::Step
-    pub(crate) fn step_failed(
+    pub(crate) fn step_failed<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         step: &gherkin::Step,
@@ -620,7 +612,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     /// [started]: event::Step::Started
     /// [`Background`]: gherkin::Background
     /// [`Step`]: gherkin::Step
-    pub(crate) fn background(
+    pub(crate) fn background<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         bg: &gherkin::Step,
@@ -641,7 +633,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
                 self.indent = self.indent.saturating_sub(4);
             }
             Step::Failed(c, w, i) => {
-                self.bg_step_failed(feat, bg, c.as_ref(), w.as_deref(), i)?;
+                self.bg_step_failed(feat, bg, c.as_ref(), w.as_ref(), i)?;
                 self.indent = self.indent.saturating_sub(4);
             }
         }
@@ -780,7 +772,7 @@ impl<W: Debug, Out: io::Write> Basic<W, Out> {
     /// [failed]: event::Step::Failed
     /// [`Background`]: gherkin::Background
     /// [`Step`]: gherkin::Step
-    pub(crate) fn bg_step_failed(
+    pub(crate) fn bg_step_failed<W: Debug>(
         &mut self,
         feat: &gherkin::Feature,
         step: &gherkin::Step,