+++ src/writer/mod.rs
@@ -12,8 +12,10 @@
 //!
 //! [`Cucumber`]: crate::event::Cucumber
 
+pub mod arbitrary_discard;
 pub mod basic;
 pub mod fail_on_skipped;
+pub mod failure_discard;
 #[cfg(feature = "output-json")]
 pub mod json;
 #[cfg(feature = "output-junit")]
@@ -28,7 +30,7 @@ use async_trait::async_trait;
 use sealed::sealed;
 use structopt::StructOptInternal;
 
-use crate::{event, parser, Event, World};
+use crate::{event, parser, Event};
 
 #[cfg(feature = "output-json")]
 #[doc(inline)]
@@ -38,8 +40,9 @@ pub use self::json::Json;
 pub use self::junit::JUnit;
 #[doc(inline)]
 pub use self::{
-    basic::Basic, fail_on_skipped::FailOnSkipped, normalized::Normalized,
-    repeat::Repeat, summarized::Summarized, tee::Tee,
+    arbitrary_discard::ArbitraryDiscard, basic::Basic,
+    fail_on_skipped::FailOnSkipped, failure_discard::FailureDiscard,
+    normalized::Normalized, repeat::Repeat, summarized::Summarized, tee::Tee,
 };
 
 /// Writer of [`Cucumber`] events to some output.
@@ -121,13 +124,12 @@ pub trait Failure<World>: Writer<World> {
 
 /// Extension of [`Writer`] allowing its normalization and summarization.
 #[sealed]
-pub trait Ext<W: World>: Sized {
+pub trait Ext: Sized {
     /// Wraps this [`Writer`] into a [`Normalized`] version.
     ///
     /// See [`Normalized`] for more information.
     #[must_use]
-    fn normalized<W: World>(self) -> Normalized<W, Self>
-    where Self: Writer<W>;
+    fn normalized<W>(self) -> Normalized<W, Self>;
 
     /// Wraps this [`Writer`] to print a summary at the end of an output.
     ///
@@ -169,7 +171,7 @@ pub trait Ext<W: World>: Sized {
     /// [`Skipped`]: event::Step::Skipped
     /// [`Step`]: gherkin::Step
     #[must_use]
-    fn repeat_skipped(self) -> Repeat<W, Self>;
+    fn repeat_skipped<W>(self) -> Repeat<W, Self>;
 
     /// Wraps this [`Writer`] to re-output [`Failed`] [`Step`]s or [`Parser`]
     /// errors at the end of an output.
@@ -178,12 +180,12 @@ pub trait Ext<W: World>: Sized {
     /// [`Parser`]: crate::Parser
     /// [`Step`]: gherkin::Step
     #[must_use]
-    fn repeat_failed(self) -> Repeat<W, Self>;
+    fn repeat_failed<W>(self) -> Repeat<W, Self>;
 
     /// Wraps this [`Writer`] to re-output `filter`ed events at the end of an
     /// output.
     #[must_use]
-    fn repeat_if<F>(self, filter: F) -> Repeat<W, Self, F>
+    fn repeat_if<W, F>(self, filter: F) -> Repeat<W, Self, F>
     where
         F: Fn(&parser::Result<Event<event::Cucumber<W>>>) -> bool;
 
@@ -193,16 +195,25 @@ pub trait Ext<W: World>: Sized {
     /// This way an event can be processed by multiple [`Writer`]s
     /// simultaneously.
     #[must_use]
-    fn tee<Wr: Writer<W>>(self, other: Wr) -> Tee<Self, Wr>;
+    fn tee<W, Wr: Writer<W>>(self, other: Wr) -> Tee<Self, Wr>;
+
+    /// Adds [`ArbitraryWriter`] implementation, which discards provided value.
+    ///
+    /// Can be useful for one of the [`Writer`]s in [`tee()`].
+    #[must_use]
+    fn discard_arbitrary(self) -> ArbitraryDiscard<Self>;
+
+    /// Adds [`FailureWriter`] implementation, which return `0` on every stat
+    /// method.
+    ///
+    /// Can be useful for one of the [`Writer`]s in [`tee()`].
+    #[must_use]
+    fn discard_failure(self) -> FailureDiscard<Self>;
 }
 
 #[sealed]
-impl<W, T> Ext<W> for T
-where
-    W: World,
-    T: Writer<W> + Sized,
-{
-    fn normalized(self) -> Normalized<W, Self> {
+impl<T> Ext for T {
+    fn normalized<W>(self) -> Normalized<W, Self> {
         Normalized::new(self)
     }
 
@@ -225,22 +236,30 @@ where
         FailOnSkipped::with(self, f)
     }
 
-    fn repeat_skipped(self) -> Repeat<W, Self> {
+    fn repeat_skipped<W>(self) -> Repeat<W, Self> {
         Repeat::skipped(self)
     }
 
-    fn repeat_failed(self) -> Repeat<W, Self> {
+    fn repeat_failed<W>(self) -> Repeat<W, Self> {
         Repeat::failed(self)
     }
 
-    fn repeat_if<F>(self, filter: F) -> Repeat<W, Self, F>
+    fn repeat_if<W, F>(self, filter: F) -> Repeat<W, Self, F>
     where
         F: Fn(&parser::Result<Event<event::Cucumber<W>>>) -> bool,
     {
         Repeat::new(self, filter)
     }
 
-    fn tee<Wr: Writer<W>>(self, other: Wr) -> Tee<Self, Wr> {
+    fn tee<W, Wr: Writer<W>>(self, other: Wr) -> Tee<Self, Wr> {
         Tee::new(self, other)
+    }
+
+    fn discard_arbitrary(self) -> ArbitraryDiscard<Self> {
+        ArbitraryDiscard::from(self)
+    }
+
+    fn discard_failure(self) -> FailureDiscard<Self> {
+        FailureDiscard::from(self)
     }
 }